“message”

+++ yo.js
@@ -96,104 +96,175 @@ app.post('/deleteUser', async (req, res) => {
 })
 
 app.post('/createUser', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   await admin.firestore().collection('Users').doc(req.body.uid).set(req.body)
   res.send({info: 'User Created'})
+  }
+  else{
+    res.send({status:400});
+  }
 })
 
 app.post('/updatePost', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   await admin.firestore().collection('Posts').doc(req.body.uid).set(req.body)
   res.send({info: 'User Updated'})
+  }
+  else{
+    res.send({status:400});
+  }
 })
 
 app.post('/getPost', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const snapshot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   res.send(snapshot.data())
+  }
+  else{
+    res.send({status:400, info: "auth failed"});
+}
 })
 
 app.post('/deletePost', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   await admin.firestore().collection('Posts').doc(req.body.uid).delete()
   res.send({info: 'Post Deleted'})
+  }
+  else{
+    res.send({status:400, info: "auth failed"});
+  }
 })
 
 app.post('/createPost', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   await admin.firestore().collection('Posts').doc(req.body.uid).set(req.body)
   res.send({info: 'Post Created'})
+  }
+  else{
+    res.send({status:400, info: "auth failed"})
+  }
 })
 
 app.post('/addLike', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const snapshot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   let postModel = snapshot.data()
   postModel.Likes += 1
   await admin.firestore().collection('Posts').doc(req.body.uid).set(postModel)
   sendNotificaiton(req.body.notiToken, req.body.title, req.body.body);
   res.send({info: "never is enough"})
+  }
+  else{
+    res.send({status:400, info: "auth failed"});
+  }
 })
 
 app.post('/unlikePost', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const snapshot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   let postModel = snapshot.data()
   postModel.Likes +- 1
   await admin.firestore().collection('Posts').doc(req.body.uid).set(postModel)
   res.send({info: "never is enough"})
+  }
+  else{
+    res.send({status:400, info: "auth failed"});
+  }
 })
 
 app.post('/addComment', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const snapshot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   let postModel = snapshot.data()
   postModel.comments.push(req.body.comment)
   await admin.firestore().collection('Posts').doc(req.body.uid).set(postModel)
   sendNotificaiton(req.body.notiToken, req.body.title, req.body.body);
   res.send({info: "never is enough"})
+  }
+  else{
+    res.send({status:400, info: "auth failed"})
+  }
 })
 
 app.post('/removeComment', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const snapshot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   let postModel = snapshot.data()
   postModel.comments.splice(postModel.comments.indexOf(req.body.comment))
   await admin.firestore().collection('Posts').doc(req.body.uid).set(postModel)
   res.send({info: "never is enough"})
+  }
+  else{
+    res.send({status:400, info: "auth failed"});
+  }
 })
 
 app.post('/addFriendRequest', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const snapshot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   let postModel = snapshot.data()
   postModel.friendRequests.push(req.body.reqUid)
   await admin.firestore().collection('Posts').doc(req.body.uid).set(postModel)
   sendNotificaiton(req.body.notiToken, req.body.title, req.body.body);
   res.send({info: "never is enough"})
+  }
+  else{
+    res.send({status:400, info: "auth failed"});
+  }
 })
 
 app.post('/declineFriendRequest', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const snapshot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   let postModel = snapshot.data()
   postModel.friendRequests.splice(postModel.friendRequests.indexOf(req.body.reqUid))
   await admin.firestore().collection('Posts').doc(req.body.uid).set(postModel)
   res.send({info: "never is enough"})
+  }
+  else{
+    res.send({status:400, info: "auth failed"});
+  }
 })
 
 app.post('/getFriendRequests', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const snapshot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   res.send(snapshot.data().friendRequests)
+  }
+  else{
+    res.send({status:400, info:"auth failed"});
+  }
 })
 
 app.post('/addFriend', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const snapshot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   let postModel = snapshot.data()
   postModel.friends.push(req.body.reqUid)
   await admin.firestore().collection('Posts').doc(req.body.uid).set(postModel)
   sendNotificaiton(req.body.notiToken, req.body.title, req.body.body);
   res.send({info: "never is enough"})
+  }
+  else{
+    res.send({status:400, info: "auth failed"});
+  }
 })
 
 app.post('/removeFriend', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const snapshot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   let postModel = snapshot.data()
   postModel.friends.splice(postModel.friends.indexOf(req.body.reqUid))
   await admin.firestore().collection('Posts').doc(req.body.uid).set(postModel)
   res.send({info: "never is enough"})
+  }
+  else{
+    res.send({status:400, info: "auth failed"});
+  }
 })
 
 app.post('/acceptFriendRequest', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const shot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   let model = shot.data()
   model.friends.push(req.body.reqUid)
@@ -205,9 +276,18 @@ app.post('/acceptFriendRequest', async (req, res) => {
   await admin.firestore().collection('Posts').doc(req.body.uid).set(postModel)
   sendNotificaiton(req.body.notiToken, req.body.title, req.body.body);
   res.send({info: "never is enough"})
+  }
+  else{
+    res.send({status:400, info: "auth failed"});
+  }
 })
 
 app.post('/getFriends', async (req, res) => {
+  if(await testAuth(req.headers['authorization'])){
   const snapshot = await admin.firestore().collection('Posts').doc(req.body.uid).get()
   res.send(snapshot.data().friends)
+  }
+  else{
+    res.send({status:400, info: "auth failed"});
+  }
 })
